backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "cognee>=0.3.6",
    "fastapi==0.116.1",
    "faststream[redis]>=0.6.6",
    "groq>=1.0.0",
    "psycopg[binary]>=3.3.3",
    "pypdf==6.0.0",
    "python-dotenv==1.1.1",
    "redis==6.4.0",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.47",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
